Refactor to be testable
